// src/proto/userPackage/User.ts
// Original file: src/proto/user.proto


export interface User {
  'id'?: (string);
  'name'?: (string);
  'createdAt'?: (string);
  'privilege'?: (string);
  'v'?: (number);
}

export interface User__Output {
  'id': (string);
  'name': (string);
  'createdAt': (string);
  'privilege': (string);
  'v': (number);
}